pyproject.toml
[project]
name = "code01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.8",
    "google-search-results>=2.4.2",
    "langchain==1.2.7",
    "langchain-anthropic>=1.3.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langchain-qianwen>=0.1.17",
    "langchain-unstructured>=1.0.1",
    "langgraph>=1.0.5",
    "loguru>=0.7.3",
    "numpy>=2.0.2",
    "ollama>=0.6.1",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "redis==5.3.1",
    "redisvl>=0.18.0",
]
